--HG--
branch : Mesh

+++ src/modules/graphics/opengl/Font.cpp
@@ -215,10 +215,10 @@ Font::Glyph *Font::addGlyph(uint32 glyph)
 		g->texture = t;
 
 		const GlyphVertex verts[4] = {
-			{0, 0, float(textureX)/float(textureWidth),   float(textureY)/float(textureHeight)},
-			{w, 0, float(textureX+w)/float(textureWidth), float(textureY)/float(textureHeight)},
-			{w, h, float(textureX+w)/float(textureWidth), float(textureY+h)/float(textureHeight)},
-			{0, h, float(textureX)/float(textureWidth),   float(textureY+h)/float(textureHeight)},
+			{    0.0f,     0.0f, float(textureX)/float(textureWidth),   float(textureY)/float(textureHeight)},
+			{float(w),     0.0f, float(textureX+w)/float(textureWidth), float(textureY)/float(textureHeight)},
+			{float(w), float(h), float(textureX+w)/float(textureWidth), float(textureY+h)/float(textureHeight)},
+			{    0.0f, float(h), float(textureX)/float(textureWidth),   float(textureY+h)/float(textureHeight)},
 		};
 
 		// copy vertex data to the glyph and set proper bearing
@@ -550,7 +550,7 @@ int Font::getDescent() const
 float Font::getBaseline() const
 {
 	// 1.25 is magic line height for true type fonts
-	return (type == FONT_TRUETYPE) ? floor(getHeight() / 1.25f + 0.5f) : 0.0f;
+	return (type == FONT_TRUETYPE) ? floorf(getHeight() / 1.25f + 0.5f) : 0.0f;
 }
 
 bool Font::hasGlyph(uint32 glyph) const
